update on structural analysis
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/cortical_surface/cortical_surface-private/trunk@39779 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ src/library/structuralanalysis/anneal.h
@@ -10,7 +10,7 @@ class Anneal: public SurfaceBased_StructuralAnalysis{
         public:
 
             void Run( int verbose = 1 );
-            void Step( std::vector<int> &random, long double temp, uint &mod );
+            void Step( std::vector<int> &random, double temp, uint &mod );
 };
